Update argument labels of functions to follow Swift 3 API Design Guidelines: https://swift.org/documentation/api-design-guidelines For example, a preposition has been split from a function name to an argument label.
Also access levels (open, public, fileprivate and private) that were automatically updated by Xcode 8 have been fixed to match their contexts. Basically a member should be `public` rather than `open` if it does not expect being overridden. A member should be `private` rather than `fileprivate` if it is inside a struct or class (not as a floating function).